+++ cmd/varlogadm/app/varlogadm.go
@@ -12,19 +12,19 @@ import (
 	"github.com/kakao/varlog/internal/varlogadm"
 )
 
-func Main(opts *varlogadm.Options) error {
+func Main(opts []varlogadm.Option, logger *zap.Logger) error {
 	// TODO: add VMSInitTimeout to options
 	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
 	defer cancel()
 
-	cm, err := varlogadm.NewClusterManager(ctx, opts)
+	cm, err := varlogadm.NewClusterManager(ctx, opts...)
 	if err != nil {
-		opts.Logger.Error("could not create cluster manager server", zap.Error(err))
+		logger.Error("could not create cluster manager server", zap.Error(err))
 		return err
 	}
 
 	if err = cm.Run(); err != nil {
-		opts.Logger.Error("could not run cluster manager server", zap.Error(err))
+		logger.Error("could not run cluster manager server", zap.Error(err))
 		return err
 	}